+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   their responses ([link to
   discussion](https://github.com/CBIIT/ccdi-federation-api/discussions/79),
   [#95](https://github.com/CBIIT/ccdi-federation-api/pull/95)).
+- Adds Library Source Material ([link to 
+  discussion](https://github.com/CBIIT/ccdi-federation-api/discussions/114))
+  [#](https://github.com/CBIIT/ccdi-federation-api/pull/))
 
 ### Revised